Extract run flow orchestration into presenter
