+++ seva/app/main.py
@@ -11,7 +11,6 @@ import time
 from datetime import datetime
 from urllib.parse import urlparse
 from tkinter import filedialog, messagebox
-from dataclasses import dataclass
 from typing import Any, Dict, Set, Optional, List, TYPE_CHECKING
 
 # ---- Views (UI-only) ----
@@ -35,21 +34,12 @@ from ..viewmodels.live_data_vm import LiveDataVM
 from ..viewmodels.runs_vm import RunsVM
 
 # ---- UseCases & Adapter ----
-from ..usecases.start_experiment_batch import StartBatchResult
-from ..usecases.run_flow_coordinator import (
-    FlowHooks,
-    FlowTick,
-    GroupContext,
-    RunFlowCoordinator,
-)
+from .run_flow_presenter import RunFlowPresenter
 from ..adapters.storage_local import StorageLocal
 from ..adapters.discovery_http import HttpDiscoveryAdapter
 from ..adapters.relay_mock import RelayMock
 from ..usecases.save_plate_layout import SavePlateLayout
 from ..usecases.load_plate_layout import LoadPlateLayout
-from ..domain.entities import ExperimentPlan
-from ..domain.plan_builder import build_meta
-from ..domain.util import well_id_to_box
 from ..domain.runs_registry import RunsRegistry
 from ..domain.ports import UseCaseError
 from ..usecases.test_relay import TestRelay
@@ -71,15 +61,6 @@ if TYPE_CHECKING:
 logging_utils.configure_root()
 
 
-@dataclass
-class FlowSession:
-    """Runtime wiring for a tracked run group managed by the app."""
-
-    coordinator: RunFlowCoordinator
-    context: GroupContext
-    after_id: Optional[str] = None
-
-
 class App:
     """Bootstrap: wire Views <-> ViewModels, REST adapter, and simple polling."""
 
@@ -94,14 +75,11 @@ class App:
             on_open_settings=self._on_open_settings,
             on_open_data_plotter=self._on_open_plotter,
         )
-        self._last_download_dir: Optional[str] = None
         self.win.bind("<Control-Shift-o>", self._on_open_download_folder_hotkey)
         self.win.bind("<Control-Shift-O>", self._on_open_download_folder_hotkey)
 
         # ---- Shared registries ----
         self.runs = RunsRegistry.instance()
-        self._sessions: Dict[str, FlowSession] = {}
-        self._active_group_id: Optional[str] = None
 
         # ---- ViewModels ----
         self.plate_vm = PlateVM(on_selection_changed=self._on_selection_changed)
@@ -174,18 +152,8 @@ class App:
         except Exception:
             self.win.tabs.add(self.runs_panel, text="Runs")
 
-        self.runs_panel.on_select = self._on_runs_select
-        self.runs_panel.on_open = self._on_runs_open_folder
-        self.runs_panel.on_cancel = self._on_runs_cancel
-        self.runs_panel.on_delete = self._on_runs_delete
-        self._refresh_runs_panel()
-
         # ---- REST Adapter & UseCases (lazy init after settings) ----
         self.controller = AppController(self.settings_vm)
-
-        # ---- Download metadata per group ----
-        self._group_storage_meta: Dict[str, Dict[str, str]] = {}
-        self._last_plan_inputs: Optional[Dict[str, Any]] = None
 
         # ---- LocalStorage Adapter ----
         self._storage_root = os.environ.get("SEVA_STORAGE_ROOT") or "."
@@ -201,7 +169,26 @@ class App:
         self.uc_set_electrode_mode = SetElectrodeMode(self._relay)
 
         # ---- Run flow coordination ----
-        self._configure_runs_registry()
+        self.run_flow = RunFlowPresenter(
+            win=self.win,
+            controller=self.controller,
+            runs=self.runs,
+            runs_vm=self.runs_vm,
+            progress_vm=self.progress_vm,
+            settings_vm=self.settings_vm,
+            storage=self._storage,
+            plate_vm=self.plate_vm,
+            experiment_vm=self.experiment_vm,
+            runs_panel=self.runs_panel,
+            ensure_adapter=self._ensure_adapter,
+            toast_error=self._toast_error,
+        )
+        self.runs_panel.on_select = self.run_flow.on_runs_select
+        self.runs_panel.on_open = self.run_flow.on_runs_open_folder
+        self.runs_panel.on_cancel = self.run_flow.on_runs_cancel
+        self.runs_panel.on_delete = self.run_flow.on_runs_delete
+        self.run_flow.refresh_runs_panel()
+        self.run_flow.configure_runs_registry(Path.home() / ".seva" / "runs_registry.json")
 
         # ---- Initial UI state (demo-ish) ----
         self._seed_demo_state()
@@ -245,181 +232,6 @@ class App:
         self.run_overview.set_boxes(boxes)
         self.activity.set_boxes(boxes)
 
-    def _configure_runs_registry(self) -> None:
-        """Configure the runs registry and re-attach persisted groups."""
-        store_path = Path.home() / ".seva" / "runs_registry.json"
-        self.runs.configure(
-            store_path=store_path,
-            hooks_factory=self._build_flow_hooks_for_group,
-            coordinator_factory=self._coordinator_factory_for_group,
-        )
-        try:
-            self.runs.load()
-        except Exception as exc:
-            self._log.warning("Failed to load runs registry: %s", exc)
-            return
-
-        for group_id in self.runs.active_groups():
-            try:
-                context = self.runs.start_tracking(group_id)
-            except Exception as exc:
-                self._log.error("Failed to re-attach group %s: %s", group_id, exc)
-                continue
-            coordinator = self.runs.coordinator_for(group_id)
-            if not coordinator or not context:
-                continue
-            self._register_session(group_id, coordinator, context)
-            self._schedule_poll(group_id, 0)
-        self._refresh_runs_panel()
-
-    def _build_flow_hooks_for_group(self, group_id: str) -> FlowHooks:
-        """Create FlowHooks bound to a specific run group."""
-        return FlowHooks(
-            on_started=lambda ctx: self._on_group_started(group_id, ctx),
-            on_snapshot=lambda snapshot: self._on_group_snapshot(group_id, snapshot),
-            on_completed=lambda path: self._on_group_completed(group_id, path),
-            on_error=lambda message: self._on_group_error(group_id, message),
-        )
-
-    def _coordinator_factory_for_group(
-        self,
-        group_id: str,
-        plan_meta: Dict[str, Any],
-        storage_meta: Dict[str, str],
-        hooks: FlowHooks,
-    ) -> RunFlowCoordinator:
-        """Factory callback passed to RunsRegistry for re-attachments."""
-        if not self._ensure_adapter():
-            raise RuntimeError("Adapters not configured for coordinator factory.")
-        coordinator = RunFlowCoordinator(
-            job_port=self.controller.job_adapter,
-            storage_port=self._storage,
-            uc_start=self.controller.uc_start,
-            uc_poll=self.controller.uc_poll,
-            uc_download=self.controller.uc_download,
-            settings=self.settings_vm,
-            hooks=hooks,
-        )
-        return coordinator
-
-    def _register_session(
-        self, group_id: str, coordinator: RunFlowCoordinator, context: GroupContext
-    ) -> None:
-        """Register a runtime session both locally and in the registry."""
-        self.runs.register_runtime(group_id, coordinator, context)
-        existing = self._sessions.get(group_id)
-        if existing and existing.after_id:
-            try:
-                self.win.after_cancel(existing.after_id)
-            except Exception:
-                pass
-        self._sessions[group_id] = FlowSession(coordinator=coordinator, context=context)
-        self._group_storage_meta.setdefault(group_id, dict(context.storage_meta))
-        if not self._active_group_id:
-            self._active_group_id = group_id
-            self.win.set_run_group_id(group_id)
-            self.runs_vm.set_active_group(group_id)
-            self.progress_vm.set_active_group(group_id, self.runs)
-        self._refresh_runs_panel()
-
-    # ==================================================================
-    # Runs panel helpers
-    # ==================================================================
-    def _refresh_runs_panel(self) -> None:
-        if not hasattr(self, "runs_panel"):
-            return
-        rows = self.runs_vm.rows()
-        self.runs_panel.set_rows(rows)
-        active = self.runs_vm.active_group_id or self._active_group_id
-        current_vm = getattr(self.progress_vm, "active_group_id", None)
-        if active and active != current_vm:
-            self.runs_panel.select_group(active)
-
-    def _open_path(self, path: str) -> None:
-        if not path:
-            return
-        try:
-            if sys.platform.startswith("win"):
-                os.startfile(path)  # type: ignore[attr-defined]
-            elif sys.platform == "darwin":
-                subprocess.check_call(["open", path])
-            else:
-                subprocess.check_call(["xdg-open", path])
-        except Exception:
-            messagebox.showwarning("Open Folder", f"Could not open folder:\n{path}")
-
-    def _on_runs_select(self, group_id: str) -> None:
-        # If already active, don't re-render.
-        if self.progress_vm.active_group_id == group_id:
-            self.runs_vm.set_active_group(group_id)
-            self._active_group_id = group_id
-            self.win.set_run_group_id(group_id)
-            return
-
-        self.runs_vm.set_active_group(group_id)
-        self.progress_vm.set_active_group(group_id, self.runs)
-        self._active_group_id = group_id
-        self.win.set_run_group_id(group_id)
-
-    def _on_runs_open_folder(self, group_id: str) -> None:
-        entry = self.runs.get(group_id)
-        if not entry:
-            return
-        path = (entry.download.path or "").strip()
-        if not path:
-            messagebox.showinfo("Open Folder", "No download directory available yet.")
-            return
-        self._open_path(path)
-
-    def _on_runs_cancel(self, group_id: str) -> None:
-        if not self._ensure_adapter() or not self.controller.uc_cancel:
-            messagebox.showinfo("Cancel Group", "Cancel use case not available.")
-            return
-
-        entry = self.runs.get(group_id)
-        if not entry:
-            messagebox.showinfo("Cancel Group", "Entry not found.")
-            return
-        if entry.status not in {"running", "pending"}:
-            messagebox.showinfo("Cancel Group", "Run is no longer active.")
-            return
-        if not messagebox.askyesno("Cancel Group", f"Really cancel group {group_id}?"):
-            return
-
-        try:
-            self.controller.uc_cancel(group_id)  # type: ignore[misc]
-            self._stop_polling(group_id)
-            self.runs.mark_cancelled(group_id)
-            self._refresh_runs_panel()
-        except Exception as exc:
-            messagebox.showerror("Cancel Group", f"Cancel failed:\n{exc}")
-
-    def _on_runs_delete(self, group_id: str) -> None:
-        entry = self.runs.get(group_id)
-        if not entry:
-            return
-
-        if entry.status in {"running", "pending"}:
-            if not messagebox.askyesno(
-                "Cancel Group", f"Group {group_id} is still running. Cancel now?"
-            ):
-                return
-            self._on_runs_cancel(group_id)
-            return
-
-        if not messagebox.askyesno("Remove", f"Remove entry {group_id} from the list?"):
-            return
-
-        self._stop_polling(group_id)
-        self.runs.remove(group_id)
-        if self.runs_vm.active_group_id == group_id:
-            self.runs_vm.set_active_group(None)
-            self.progress_vm.set_active_group(None, self.runs)
-        if self._active_group_id == group_id:
-            self._active_group_id = next(iter(self._sessions), None)
-            self.win.set_run_group_id(self._active_group_id or "")
-        self._refresh_runs_panel()
-
     # ==================================================================
     # Adapter wiring
     # ==================================================================
@@ -448,177 +260,13 @@ class App:
     # ==================================================================
     def _on_submit(self) -> None:
         """Handle toolbar submit triggered by the user."""
-        # Submit: kick off coordinator flow.
-        try:
-            if not self._ensure_adapter():
-                return
-
-            configured = self.plate_vm.configured()
-            if not configured:
-                self.win.show_toast("No configured wells to start.")
-                return
-
-            selection = self.plate_vm.get_selection()
-            self.experiment_vm.set_selection(selection)
-            plan = self._build_domain_plan()
-            boxes = sorted(
-                {
-                    box
-                    for wid in configured
-                    if isinstance(wid, str)
-                    for box in [well_id_to_box(wid)]
-                    if box is not None
-                }
-            )
-            summary = {
-                "wells": len(configured),
-                "boxes": boxes or ["-"],
-                "stream": bool(self.settings_vm.use_streaming),
-            }
-            self._log.info("Submitting start request: %s", summary)
-            self._log.debug("Start selection=%s", sorted(configured))
-
-            start_hooks = FlowHooks()
-            coordinator = RunFlowCoordinator(
-                job_port=self.controller.job_adapter,
-                storage_port=self._storage,
-                uc_start=self.controller.uc_start,
-                uc_poll=self.controller.uc_poll,
-                uc_download=self.controller.uc_download,
-                settings=self.settings_vm,
-                hooks=start_hooks,
-            )
-
-            try:
-                ctx = coordinator.start(plan)
-            except UseCaseError as e:
-                if getattr(e, "code", "") == "SLOT_BUSY":
-                    meta = getattr(e, "meta", None) or {}
-                    busy = []
-                    if isinstance(meta, dict) and isinstance(meta.get("busy_wells"), list):
-                        busy = [str(w) for w in meta.get("busy_wells")]
-                    msg = e.message or "Slots busy."
-                    try:
-                        # If MainWindowView supports a warn kind, otherwise plain toast
-                        self.win.show_toast(f"Start abgelehnt: {msg}")
-                    except Exception:
-                        self.win.show_toast(f"Start abgelehnt: {msg}")
-                    # Optional: flash wells if supported by PlateVM view
-                    if busy and hasattr(self.plate_vm, "flash_wells"):
-                        try:
-                            self.plate_vm.flash_wells(busy)
-                        except Exception:
-                            pass
-                    coordinator.stop_polling()
-                    return
-                raise
-            start_result = coordinator.last_start_result()
-            if not isinstance(start_result, StartBatchResult):
-                raise RuntimeError("Coordinator returned an unexpected start result.")
-
-            if not start_result.run_group_id:
-                coordinator.stop_polling()
-                return
-
-            group_id = str(start_result.run_group_id)
-            subruns = start_result.per_box_runs
-            meta = plan.meta
-            storage_inputs = self._last_plan_inputs or {}
-            normalized_storage = {
-                "experiment": meta.experiment,
-                "subdir": meta.subdir or "",
-                "client_datetime": self._format_client_datetime_for_storage(meta.client_dt.value),
-                "results_dir": str(
-                    storage_inputs.get("results_dir") or self.settings_vm.results_dir or ""
-                ).strip()
-                or self.settings_vm.results_dir,
-            }
-            self._group_storage_meta[group_id] = normalized_storage
-            self._log.info(
-                "Start response: group=%s wells=%d boxes=%s",
-                group_id,
-                len(start_result.started_wells),
-                sorted(subruns.keys()),
-            )
-            self._log.debug("Start run map: %s", subruns)
-
-            ctx.storage_meta.update(normalized_storage)
-            plan_meta_payload = {
-                "experiment": meta.experiment,
-                "subdir": meta.subdir or "",
-                "client_datetime": meta.client_dt.value.isoformat(),
-                "group_id": group_id,
-            }
-            self.runs.add(
-                group_id=group_id,
-                name=meta.experiment,
-                boxes=sorted(subruns.keys()),
-                runs_by_box=subruns,
-                plan_meta=plan_meta_payload,
-                storage_meta=normalized_storage,
-            )
-            self._refresh_runs_panel()
-
-            tracking_hooks = self._build_flow_hooks_for_group(group_id)
-            coordinator.hooks = tracking_hooks
-            self._register_session(group_id, coordinator, ctx)
-            self.runs.start_tracking(group_id)
-
-            self._active_group_id = group_id
-            self.win.set_run_group_id(group_id)
-
-            started_boxes = ", ".join(sorted(subruns.keys()))
-            if started_boxes:
-                self.win.show_toast(f"Started group {group_id} on {started_boxes}")
-            else:
-                self.win.show_toast(f"Started group {group_id}.")
-            self._schedule_poll(group_id, 0)
-        except Exception as e:
-            self._stop_polling()
-            self._toast_error(e)
+        self.run_flow.start_run()
 
     def _on_cancel_group(self) -> None:
-        if not self._active_group_id or not self._ensure_adapter():
-            self.win.show_toast("No active group.")
-            return
-        try:
-            current = self._active_group_id
-            self._log.info("Cancel requested for group %s", current)
-            self.controller.uc_cancel(current)  # type: ignore[misc]
-            self._stop_polling(current)
-            self.runs.mark_cancelled(current)
-            self.win.show_toast(f"Cancel requested for group {current}.")
-            self._refresh_runs_panel()
-            self.progress_vm.set_active_group(current, self.runs)
-        except Exception as e:
-            self._toast_error(e)
+        self.run_flow.cancel_active_group()
 
     def _on_end_selection(self) -> None:
-        selection = sorted(self.plate_vm.get_selection())
-        if not selection:
-            self.win.show_toast("Select at least one well.")
-            return
-        if not self._ensure_adapter():
-            return
-        cancel_runs = self.controller.uc_cancel_runs
-        if cancel_runs is None:
-            self.win.show_toast("Cancel selected runs not available.")
-            return
-
-        payload = self.progress_vm.map_selection_to_runs(selection)
-        if not payload:
-            self.win.show_toast("Selected wells have no active runs.")
-            return
-
-        try:
-            self._log.info(
-                "End selection requested for wells: %s", ", ".join(selection)
-            )
-            request = {"box_runs": payload, "span": "selected"}
-            cancel_runs(request)
-            self.win.show_toast("Abort requested for selected runs.")
-        except Exception as exc:
-            self._toast_error(exc, context="Cancel runs")
+        self.run_flow.cancel_selected_runs()
 
     def _suggest_layout_filename(self) -> str:
         """Return a default layout filename derived from the experiment name."""
@@ -1103,7 +751,7 @@ class App:
             return
 
         # reset adapter to reflect new settings
-        self._stop_polling()
+        self.run_flow.stop_all_polling()
         self.controller.reset()
         self._apply_box_configuration()
         self.win.show_toast("Settings saved.")
@@ -1112,11 +760,11 @@ class App:
         DataProcessingGUI(self.win)
 
     def _on_download_group_results(self) -> None:
-        group_id = self._active_group_id
+        group_id = self.run_flow.active_group_id
         if not group_id or not self._ensure_adapter():
             self.win.show_toast("No active group.")
             return
-        storage_meta = self._group_storage_meta.get(group_id)
+        storage_meta = self.run_flow.group_storage_meta_for(group_id)
         if not storage_meta:
             self.win.show_toast(
                 "Missing storage metadata for the active group. Start must finish before downloading."
@@ -1137,41 +785,14 @@ class App:
                 "Downloaded group %s to %s", group_id, out_dir
             )
             resolved_dir = os.path.abspath(out_dir)
-            self._last_download_dir = resolved_dir
-            self.win.show_toast(self._build_download_toast(group_id, resolved_dir))
+            self.run_flow.record_download_dir(resolved_dir)
+            self.win.show_toast(self.run_flow.build_download_toast(group_id, resolved_dir))
         except Exception as e:
             self._toast_error(e)
 
     def _on_download_box_results(self, box_id: str) -> None:
         # Group ZIPs are per group; per-box filtering could be added in adapter if needed
         self._on_download_group_results()
-
-    def _build_download_toast(self, group_id: str, path: str) -> str:
-        short_path = self._shorten_download_path(path)
-        descriptor = ""
-        meta = self._group_storage_meta.get(group_id) if group_id else None
-        if not meta and group_id:
-            entry = self.runs.get(group_id)
-            if entry:
-                meta = entry.storage_meta
-        if meta:
-            parts = [
-                str(meta.get("experiment") or "").strip(),
-                str(meta.get("subdir") or "").strip(),
-                str(meta.get("client_datetime") or "").strip(),
-            ]
-            descriptor = "/".join([p for p in parts if p])
-        target = f"{descriptor} → {short_path}" if descriptor else short_path
-        if self._can_open_results_folder():
-            return f"Results unpacked to {target} (Ctrl+Shift+O to open)"
-        return f"Results unpacked to {target}"
-
-    def _shorten_download_path(self, path: str, max_len: int = 60) -> str:
-        normalized = os.path.normpath(path)
-        if len(normalized) <= max_len:
-            return normalized
-        suffix_len = max(3, max_len - 3)
-        return f"...{normalized[-suffix_len:]}"
 
     def _can_open_results_folder(self) -> bool:
         if sys.platform.startswith("win"):
@@ -1181,13 +802,13 @@ class App:
         return shutil.which("xdg-open") is not None
 
     def _on_open_download_folder_hotkey(self, event=None):
-        if not self._last_download_dir:
+        if not self.run_flow.last_download_dir:
             self.win.show_toast("Nothing downloaded yet.")
             return "break"
         if not self._can_open_results_folder():
             self.win.show_toast("Open folder not supported on this platform.")
             return "break"
-        self._open_results_folder(self._last_download_dir)
+        self._open_results_folder(self.run_flow.last_download_dir)
         return "break"
 
     def _open_results_folder(self, path: str) -> None:
@@ -1347,151 +968,6 @@ class App:
         self.win.show_toast(f"Open PNG for {well_id}")
 
     # ==================================================================
-    # Polling helpers
-    # ==================================================================
-    def _stop_polling(self, group_id: Optional[str] = None) -> None:
-        """Cancel scheduled polls and stop coordinators."""
-        if group_id is None:
-            for gid in list(self._sessions.keys()):
-                self._stop_polling(gid)
-            return
-
-        session = self._sessions.get(group_id)
-        if not session:
-            return
-        self._cancel_poll_timer(group_id)
-        try:
-            session.coordinator.stop_polling()
-        except Exception:
-            pass
-        self.runs.unregister_runtime(group_id)
-        self._sessions.pop(group_id, None)
-        if self._active_group_id == group_id:
-            self._active_group_id = next(iter(self._sessions), None)
-            self.win.set_run_group_id(self._active_group_id or "")
-
-    def _cancel_poll_timer(self, group_id: Optional[str]) -> None:
-        """Cancel scheduled Tk timer(s) for polling."""
-        if group_id is None:
-            for gid in list(self._sessions.keys()):
-                self._cancel_poll_timer(gid)
-            return
-        session = self._sessions.get(group_id)
-        if session and session.after_id is not None:
-            try:
-                self.win.after_cancel(session.after_id)
-            except Exception:
-                pass
-            session.after_id = None
-
-    def _schedule_poll(self, group_id: str, delay_ms: int) -> None:
-        """Schedule the next poll tick for a given group."""
-        session = self._sessions.get(group_id)
-        if not session:
-            return
-        delay = max(1, int(delay_ms))
-        self._cancel_poll_timer(group_id)
-        session.after_id = self.win.after(
-            delay, lambda gid=group_id: self._on_poll_tick(gid)
-        )
-
-    def _on_poll_tick(self, group_id: str) -> None:
-        """Cooperative poll tick executed on the Tkinter thread."""
-        session = self._sessions.get(group_id)
-        if not session:
-            return
-        if not self._ensure_adapter():
-            return
-
-        coordinator = session.coordinator
-        context = session.context
-
-        tick: FlowTick = coordinator.poll_once(context)
-        if tick.event == "tick":
-            delay = tick.next_delay_ms
-            if delay is None:
-                base = getattr(self.settings_vm, "poll_interval_ms", 1000) or 1000
-                try:
-                    delay = max(200, int(base))
-                except (TypeError, ValueError):
-                    delay = 1000
-            self._schedule_poll(group_id, int(delay))
-            return
-
-        # Completed/Error: no further scheduling; rely on hooks for UI feedback.
-        self._cancel_poll_timer(group_id)
-        if tick.event == "completed":
-            coordinator.stop_polling()
-            auto_download_enabled = getattr(
-                self.settings_vm, "auto_download_on_complete", True
-            )
-            download_path: Optional[Path] = None
-            if tick.snapshot:
-                try:
-                    download_path = coordinator.on_completed(context, tick.snapshot)
-                except Exception as exc:
-                    self._toast_error(exc, context="Download failed")
-                    self._finalize_session(group_id)
-                    return
-            if not auto_download_enabled:
-                self._on_group_completed(group_id, None)
-            self._finalize_session(group_id)
-            return
-
-        if tick.event == "error":
-            coordinator.stop_polling()
-            self._finalize_session(group_id)
-
-    def _finalize_session(self, group_id: str) -> None:
-        """Clean up local bookkeeping once a run no longer needs polling."""
-        session = self._sessions.pop(group_id, None)
-        if session and session.after_id is not None:
-            try:
-                self.win.after_cancel(session.after_id)
-            except Exception:
-                pass
-        self.runs.unregister_runtime(group_id)
-        if self._active_group_id == group_id:
-            self._active_group_id = next(iter(self._sessions), None)
-            self.win.set_run_group_id(self._active_group_id or "")
-        self._refresh_runs_panel()
-
-    def _on_group_started(self, group_id: str, ctx: GroupContext) -> None:
-        """Hook fired when a coordinator acknowledges start for a group."""
-        self._log.debug("Coordinator acknowledged start for group %s", ctx.group)
-
-    def _on_group_snapshot(self, group_id: str, snapshot) -> None:
-        """Hook fired on every polling snapshot for a group."""
-        self.runs.update_snapshot(group_id, snapshot)
-        if snapshot and group_id == self._active_group_id:
-            self.progress_vm.apply_snapshot(snapshot)
-        self._refresh_runs_panel()
-
-    def _on_group_completed(self, group_id: str, path: Optional[Path]) -> None:
-        """Hook fired when the flow reports completion for a group."""
-        download_path = os.path.abspath(str(path)) if path else None
-        self.runs.mark_done(group_id, download_path)
-        if download_path:
-            self._last_download_dir = download_path
-            self.win.show_toast(self._build_download_toast(group_id, download_path))
-        else:
-            self.win.show_toast("All runs completed.")
-        self._refresh_runs_panel()
-        if getattr(self.progress_vm, "active_group_id", None) == group_id:
-            self.progress_vm.set_active_group(group_id, self.runs)
-
-    def _on_group_error(self, group_id: str, message: str) -> None:
-        """Hook fired when the flow reports a polling error for a group."""
-        text = message.strip() if isinstance(message, str) else ""
-        if text:
-            self._log.error("Polling error for %s: %s", group_id, text)
-        else:
-            self._log.error("Polling error for %s: <empty message>", group_id)
-        self.runs.mark_error(group_id, text or None)
-        self.win.show_toast(text or "Polling failed.")
-        self._refresh_runs_panel()
-
-    # ==================================================================
     # Error handling helpers
     # ==================================================================
 
@@ -1571,99 +1047,7 @@ class App:
                     return slot
         return None
 
-    # ==================================================================
-    # Plan building
-    # ==================================================================
-
-    def _build_domain_plan(self) -> ExperimentPlan:
-        """Build a domain ExperimentPlan from the current view-model state."""
-        configured = self.plate_vm.configured()
-        if not configured:
-            raise RuntimeError("No configured wells to start.")
-        
-        well_plan_list = self.experiment_vm.build_well_plan_map(configured)
-        if not well_plan_list:
-            raise RuntimeError("No saved parameters found for configured wells.")
-
-        inputs = self._collect_plan_inputs()
-        meta = build_meta(
-            experiment=inputs["experiment"],
-            subdir=inputs["subdir"],
-            client_dt_local=inputs["client_dt"],
-        )
-        # Persist plan inputs so downstream UI storage metadata can reuse them.
-        self._last_plan_inputs = inputs
-
-        return ExperimentPlan(
-            meta=meta,
-            wells=well_plan_list
-        )
-
-    def _collect_plan_inputs(self) -> Dict[str, Any]:
-        """Gather experiment metadata and filesystem settings for plan construction."""
-        experiment_name = (self.settings_vm.experiment_name or "").strip()
-        if not experiment_name:
-            raise RuntimeError("Experiment name must be set in Settings.")
-
-        subdir_raw = (self.settings_vm.subdir or "").strip()
-        subdir = subdir_raw or None
-
-        override_dt = ""
-        if hasattr(self.experiment_vm, "fields"):
-            override_dt = str(self.experiment_vm.fields.get("storage.client_datetime") or "")
-        client_dt = self._parse_client_datetime_override(override_dt)
-
-        results_dir = str(self.settings_vm.results_dir or "").strip() or "."
-
-        return {
-            "experiment": experiment_name,
-            "subdir": subdir,
-            "client_dt": client_dt,
-            "results_dir": results_dir,
-        }
-
-    def _parse_client_datetime_override(self, value: str) -> datetime:
-        """Interpret stored client datetime overrides, falling back to the current time."""
-        text = str(value or "").strip()
-        if not text:
-            return datetime.now().astimezone().replace(microsecond=0)
-
-        normalized = text.replace(" ", "T")
-        if normalized.endswith("Z"):
-            normalized = normalized[:-1] + "+00:00"
-        if "T" in normalized:
-            date_part, time_part = normalized.split("T", 1)
-            if ":" not in time_part:
-                time_part = time_part.replace("-", ":")
-            normalized = f"{date_part}T{time_part}"
-
-        parsed = None
-        try:
-            parsed = datetime.fromisoformat(normalized)
-        except ValueError:
-            pass
-
-        if parsed is None:
-            for fmt in ("%Y-%m-%d_%H-%M-%S", "%Y-%m-%d %H:%M:%S", "%Y-%m-%dT%H-%M-%S"):
-                try:
-                    parsed = datetime.strptime(text, fmt)
-                    break
-                except ValueError:
-                    continue
-
-        if parsed is None:
-            return datetime.now().astimezone().replace(microsecond=0)
-
-        if parsed.tzinfo is None or parsed.tzinfo.utcoffset(parsed) is None:
-            local_zone = datetime.now().astimezone().tzinfo
-            parsed = parsed.replace(tzinfo=local_zone)
-
-        return parsed.astimezone().replace(microsecond=0)
-
-    @staticmethod
-    def _format_client_datetime_for_storage(dt: datetime) -> str:
-        """Return a filesystem-safe representation of the client datetime."""
-        return dt.astimezone().strftime("%Y-%m-%d_%H-%M-%S")
+    
 
 
 def main() -> None:
