+++ seva/app/run_flow_presenter.py
@@ -1,20 +1,716 @@
 from __future__ import annotations
 
-"""UI-facing presenter that coordinates run flow actions without view logic.
+"""UI-facing presenter that coordinates run flow actions without view logic."""
 
-Inputs:
-- callbacks or UI adapters for start/cancel/poll notifications
-- use cases and coordinators for run flow orchestration
+import logging
+import os
+import shutil
+import subprocess
+import sys
+from dataclasses import dataclass
+from datetime import datetime
+from pathlib import Path
+from typing import Any, Dict, Optional
+from tkinter import messagebox
 
-Outputs:
-- invokes hooks/callbacks when run state changes
-- returns run identifiers and context data to the caller
-"""
+from seva.domain.entities import ExperimentPlan
+from seva.domain.ports import UseCaseError
+from seva.domain.util import well_id_to_box
+from seva.domain.runs_registry import RunsRegistry
+from seva.domain.plan_builder import build_meta
+from seva.usecases.run_flow_coordinator import (
+    FlowHooks,
+    FlowTick,
+    GroupContext,
+    RunFlowCoordinator,
+)
+from seva.usecases.start_experiment_batch import StartBatchResult
+from seva.viewmodels.experiment_vm import ExperimentVM
+from seva.viewmodels.plate_vm import PlateVM
+from seva.viewmodels.progress_vm import ProgressVM
+from seva.viewmodels.runs_vm import RunsVM
+from seva.viewmodels.settings_vm import SettingsVM
+from seva.app.polling_scheduler import PollingScheduler
+
+from .controller import AppController
+from ..adapters.storage_local import StorageLocal
+
+
+@dataclass
+class FlowSession:
+    """Runtime wiring for a tracked run group managed by the app."""
+
+    coordinator: RunFlowCoordinator
+    context: GroupContext
 
 
 class RunFlowPresenter:
-    """Placeholder for run-flow orchestration extracted from main.py."""
+    """Coordinates start/cancel/poll logic and run registry updates."""
 
-    def __init__(self, *args, **kwargs) -> None:
-        self._args = args
-        self._kwargs = kwargs
+    def __init__(
+        self,
+        *,
+        win,
+        controller: AppController,
+        runs: RunsRegistry,
+        runs_vm: RunsVM,
+        progress_vm: ProgressVM,
+        settings_vm: SettingsVM,
+        storage: StorageLocal,
+        plate_vm: PlateVM,
+        experiment_vm: ExperimentVM,
+        runs_panel,
+        ensure_adapter,
+        toast_error,
+    ) -> None:
+        self._log = logging.getLogger(__name__)
+        self.win = win
+        self.controller = controller
+        self.runs = runs
+        self.runs_vm = runs_vm
+        self.progress_vm = progress_vm
+        self.settings_vm = settings_vm
+        self.storage = storage
+        self.plate_vm = plate_vm
+        self.experiment_vm = experiment_vm
+        self.runs_panel = runs_panel
+        self._ensure_adapter = ensure_adapter
+        self._toast_error = toast_error
+
+        self._sessions: Dict[str, FlowSession] = {}
+        self._active_group_id: Optional[str] = None
+        self._group_storage_meta: Dict[str, Dict[str, str]] = {}
+        self._last_plan_inputs: Optional[Dict[str, Any]] = None
+        self._last_download_dir: Optional[str] = None
+
+        self._scheduler = PollingScheduler(win.after, win.after_cancel)
+
+    # ------------------------------------------------------------------
+    # Public properties
+    # ------------------------------------------------------------------
+    @property
+    def active_group_id(self) -> Optional[str]:
+        return self._active_group_id
+
+    @property
+    def last_download_dir(self) -> Optional[str]:
+        return self._last_download_dir
+
+    def group_storage_meta_for(self, group_id: str) -> Optional[Dict[str, str]]:
+        return self._group_storage_meta.get(group_id)
+
+    def record_download_dir(self, path: str) -> None:
+        self._last_download_dir = path
+
+    # ------------------------------------------------------------------
+    # Registry wiring
+    # ------------------------------------------------------------------
+    def configure_runs_registry(self, store_path: Path) -> None:
+        """Configure the runs registry and re-attach persisted groups."""
+        self.runs.configure(
+            store_path=store_path,
+            hooks_factory=self._build_flow_hooks_for_group,
+            coordinator_factory=self._coordinator_factory_for_group,
+        )
+        try:
+            self.runs.load()
+        except Exception as exc:
+            self._log.warning("Failed to load runs registry: %s", exc)
+            return
+
+        for group_id in self.runs.active_groups():
+            try:
+                context = self.runs.start_tracking(group_id)
+            except Exception as exc:
+                self._log.error("Failed to re-attach group %s: %s", group_id, exc)
+                continue
+            coordinator = self.runs.coordinator_for(group_id)
+            if not coordinator or not context:
+                continue
+            self._register_session(group_id, coordinator, context)
+            self._schedule_poll(group_id, 0)
+        self._refresh_runs_panel()
+
+    def _build_flow_hooks_for_group(self, group_id: str) -> FlowHooks:
+        """Create FlowHooks bound to a specific run group."""
+        return FlowHooks(
+            on_started=lambda ctx: self._on_group_started(group_id, ctx),
+            on_snapshot=lambda snapshot: self._on_group_snapshot(group_id, snapshot),
+            on_completed=lambda path: self._on_group_completed(group_id, path),
+            on_error=lambda message: self._on_group_error(group_id, message),
+        )
+
+    def _coordinator_factory_for_group(
+        self,
+        group_id: str,
+        plan_meta: Dict[str, Any],
+        storage_meta: Dict[str, str],
+        hooks: FlowHooks,
+    ) -> RunFlowCoordinator:
+        """Factory callback passed to RunsRegistry for re-attachments."""
+        if not self._ensure_adapter():
+            raise RuntimeError("Adapters not configured for coordinator factory.")
+        coordinator = RunFlowCoordinator(
+            job_port=self.controller.job_adapter,
+            storage_port=self.storage,
+            uc_start=self.controller.uc_start,
+            uc_poll=self.controller.uc_poll,
+            uc_download=self.controller.uc_download,
+            settings=self.settings_vm,
+            hooks=hooks,
+        )
+        return coordinator
+
+    def _register_session(
+        self, group_id: str, coordinator: RunFlowCoordinator, context: GroupContext
+    ) -> None:
+        """Register a runtime session both locally and in the registry."""
+        self.runs.register_runtime(group_id, coordinator, context)
+        self._sessions[group_id] = FlowSession(coordinator=coordinator, context=context)
+        self._group_storage_meta.setdefault(group_id, dict(context.storage_meta))
+        if not self._active_group_id:
+            self._set_active_group(group_id)
+        self._refresh_runs_panel()
+
+    # ------------------------------------------------------------------
+    # Runs panel helpers
+    # ------------------------------------------------------------------
+    def _refresh_runs_panel(self) -> None:
+        if not self.runs_panel:
+            return
+        rows = self.runs_vm.rows()
+        self.runs_panel.set_rows(rows)
+        active = self.runs_vm.active_group_id or self._active_group_id
+        current_vm = getattr(self.progress_vm, "active_group_id", None)
+        if active and active != current_vm:
+            self.runs_panel.select_group(active)
+
+    def refresh_runs_panel(self) -> None:
+        self._refresh_runs_panel()
+
+    def build_download_toast(self, group_id: str, path: str) -> str:
+        return self._build_download_toast(group_id, path)
+
+    def stop_all_polling(self) -> None:
+        self._stop_polling()
+
+    def _open_path(self, path: str) -> None:
+        if not path:
+            return
+        try:
+            if sys.platform.startswith("win"):
+                os.startfile(path)  # type: ignore[attr-defined]
+            elif sys.platform == "darwin":
+                subprocess.check_call(["open", path])
+            else:
+                subprocess.check_call(["xdg-open", path])
+        except Exception:
+            messagebox.showwarning("Open Folder", f"Could not open folder:\n{path}")
+
+    def on_runs_select(self, group_id: str) -> None:
+        if self.progress_vm.active_group_id == group_id:
+            self.runs_vm.set_active_group(group_id)
+            self._active_group_id = group_id
+            self.win.set_run_group_id(group_id)
+            return
+
+        self.runs_vm.set_active_group(group_id)
+        self.progress_vm.set_active_group(group_id, self.runs)
+        self._active_group_id = group_id
+        self.win.set_run_group_id(group_id)
+
+    def on_runs_open_folder(self, group_id: str) -> None:
+        entry = self.runs.get(group_id)
+        if not entry:
+            return
+        path = (entry.download.path or "").strip()
+        if not path:
+            messagebox.showinfo("Open Folder", "No download directory available yet.")
+            return
+        self._open_path(path)
+
+    def on_runs_cancel(self, group_id: str) -> None:
+        if not self._ensure_adapter() or not self.controller.uc_cancel:
+            messagebox.showinfo("Cancel Group", "Cancel use case not available.")
+            return
+
+        entry = self.runs.get(group_id)
+        if not entry:
+            messagebox.showinfo("Cancel Group", "Entry not found.")
+            return
+        if entry.status not in {"running", "pending"}:
+            messagebox.showinfo("Cancel Group", "Run is no longer active.")
+            return
+        if not messagebox.askyesno("Cancel Group", f"Really cancel group {group_id}?"):
+            return
+
+        try:
+            self.controller.uc_cancel(group_id)  # type: ignore[misc]
+            self._stop_polling(group_id)
+            self.runs.mark_cancelled(group_id)
+            self._refresh_runs_panel()
+        except Exception as exc:
+            messagebox.showerror("Cancel Group", f"Cancel failed:\n{exc}")
+
+    def on_runs_delete(self, group_id: str) -> None:
+        entry = self.runs.get(group_id)
+        if not entry:
+            return
+
+        if entry.status in {"running", "pending"}:
+            if not messagebox.askyesno(
+                "Cancel Group", f"Group {group_id} is still running. Cancel now?"
+            ):
+                return
+            self.on_runs_cancel(group_id)
+            return
+
+        if not messagebox.askyesno("Remove", f"Remove entry {group_id} from the list?"):
+            return
+
+        self._stop_polling(group_id)
+        self.runs.remove(group_id)
+        if self.runs_vm.active_group_id == group_id:
+            self.runs_vm.set_active_group(None)
+            self.progress_vm.set_active_group(None, self.runs)
+        if self._active_group_id == group_id:
+            self._active_group_id = next(iter(self._sessions), None)
+            self.win.set_run_group_id(self._active_group_id or "")
+        self._refresh_runs_panel()
+
+    # ------------------------------------------------------------------
+    # Run flow actions
+    # ------------------------------------------------------------------
+    def start_run(self) -> None:
+        """Handle toolbar submit triggered by the user."""
+        try:
+            if not self._ensure_adapter():
+                return
+
+            configured = self.plate_vm.configured()
+            if not configured:
+                self.win.show_toast("No configured wells to start.")
+                return
+
+            selection = self.plate_vm.get_selection()
+            self.experiment_vm.set_selection(selection)
+            plan = self._build_domain_plan()
+            boxes = sorted(
+                {
+                    box
+                    for wid in configured
+                    if isinstance(wid, str)
+                    for box in [well_id_to_box(wid)]
+                    if box is not None
+                }
+            )
+            summary = {
+                "wells": len(configured),
+                "boxes": boxes or ["-"],
+                "stream": bool(self.settings_vm.use_streaming),
+            }
+            self._log.info("Submitting start request: %s", summary)
+            self._log.debug("Start selection=%s", sorted(configured))
+
+            start_hooks = FlowHooks()
+            coordinator = RunFlowCoordinator(
+                job_port=self.controller.job_adapter,
+                storage_port=self.storage,
+                uc_start=self.controller.uc_start,
+                uc_poll=self.controller.uc_poll,
+                uc_download=self.controller.uc_download,
+                settings=self.settings_vm,
+                hooks=start_hooks,
+            )
+
+            try:
+                ctx = coordinator.start(plan)
+            except UseCaseError as exc:
+                if getattr(exc, "code", "") == "SLOT_BUSY":
+                    meta = getattr(exc, "meta", None) or {}
+                    busy = []
+                    if isinstance(meta, dict) and isinstance(meta.get("busy_wells"), list):
+                        busy = [str(w) for w in meta.get("busy_wells")]
+                    msg = exc.message or "Slots busy."
+                    try:
+                        self.win.show_toast(f"Start rejected: {msg}")
+                    except Exception:
+                        self.win.show_toast(f"Start rejected: {msg}")
+                    if busy and hasattr(self.plate_vm, "flash_wells"):
+                        try:
+                            self.plate_vm.flash_wells(busy)
+                        except Exception:
+                            pass
+                    coordinator.stop_polling()
+                    return
+                raise
+            start_result = coordinator.last_start_result()
+            if not isinstance(start_result, StartBatchResult):
+                raise RuntimeError("Coordinator returned an unexpected start result.")
+
+            if not start_result.run_group_id:
+                coordinator.stop_polling()
+                return
+
+            group_id = str(start_result.run_group_id)
+            subruns = start_result.per_box_runs
+            meta = plan.meta
+            storage_inputs = self._last_plan_inputs or {}
+            normalized_storage = {
+                "experiment": meta.experiment,
+                "subdir": meta.subdir or "",
+                "client_datetime": self._format_client_datetime_for_storage(meta.client_dt.value),
+                "results_dir": str(
+                    storage_inputs.get("results_dir") or self.settings_vm.results_dir or ""
+                ).strip()
+                or self.settings_vm.results_dir,
+            }
+            self._group_storage_meta[group_id] = normalized_storage
+            self._log.info(
+                "Start response: group=%s wells=%d boxes=%s",
+                group_id,
+                len(start_result.started_wells),
+                sorted(subruns.keys()),
+            )
+            self._log.debug("Start run map: %s", subruns)
+
+            ctx.storage_meta.update(normalized_storage)
+            plan_meta_payload = {
+                "experiment": meta.experiment,
+                "subdir": meta.subdir or "",
+                "client_datetime": meta.client_dt.value.isoformat(),
+                "group_id": group_id,
+            }
+            self.runs.add(
+                group_id=group_id,
+                name=meta.experiment,
+                boxes=sorted(subruns.keys()),
+                runs_by_box=subruns,
+                plan_meta=plan_meta_payload,
+                storage_meta=normalized_storage,
+            )
+            self._refresh_runs_panel()
+
+            tracking_hooks = self._build_flow_hooks_for_group(group_id)
+            coordinator.hooks = tracking_hooks
+            self._register_session(group_id, coordinator, ctx)
+            self.runs.start_tracking(group_id)
+
+            self._set_active_group(group_id)
+
+            started_boxes = ", ".join(sorted(subruns.keys()))
+            if started_boxes:
+                self.win.show_toast(f"Started group {group_id} on {started_boxes}")
+            else:
+                self.win.show_toast(f"Started group {group_id}.")
+            self._schedule_poll(group_id, 0)
+        except Exception as exc:
+            self._stop_polling()
+            self._toast_error(exc)
+
+    def cancel_active_group(self) -> None:
+        if not self._active_group_id or not self._ensure_adapter():
+            self.win.show_toast("No active group.")
+            return
+        try:
+            current = self._active_group_id
+            self._log.info("Cancel requested for group %s", current)
+            self.controller.uc_cancel(current)  # type: ignore[misc]
+            self._stop_polling(current)
+            self.runs.mark_cancelled(current)
+            self.win.show_toast(f"Cancel requested for group {current}.")
+            self._refresh_runs_panel()
+            self.progress_vm.set_active_group(current, self.runs)
+        except Exception as exc:
+            self._toast_error(exc)
+
+    def cancel_selected_runs(self) -> None:
+        selection = sorted(self.plate_vm.get_selection())
+        if not selection:
+            self.win.show_toast("Select at least one well.")
+            return
+        if not self._ensure_adapter():
+            return
+        cancel_runs = self.controller.uc_cancel_runs
+        if cancel_runs is None:
+            self.win.show_toast("Cancel selected runs not available.")
+            return
+
+        payload = self.progress_vm.map_selection_to_runs(selection)
+        if not payload:
+            self.win.show_toast("Selected wells have no active runs.")
+            return
+
+        try:
+            self._log.info(
+                "End selection requested for wells: %s", ", ".join(selection)
+            )
+            request = {"box_runs": payload, "span": "selected"}
+            cancel_runs(request)
+            self.win.show_toast("Abort requested for selected runs.")
+        except Exception as exc:
+            self._toast_error(exc, context="Cancel runs")
+
+    # ------------------------------------------------------------------
+    # Polling helpers
+    # ------------------------------------------------------------------
+    def _stop_polling(self, group_id: Optional[str] = None) -> None:
+        """Cancel scheduled polls and stop coordinators."""
+        if group_id is None:
+            for gid in list(self._sessions.keys()):
+                self._stop_polling(gid)
+            return
+
+        session = self._sessions.get(group_id)
+        if not session:
+            return
+        self._scheduler.cancel(group_id)
+        try:
+            session.coordinator.stop_polling()
+        except Exception:
+            pass
+        self.runs.unregister_runtime(group_id)
+        self._sessions.pop(group_id, None)
+        if self._active_group_id == group_id:
+            self._active_group_id = next(iter(self._sessions), None)
+            self.win.set_run_group_id(self._active_group_id or "")
+
+    def _schedule_poll(self, group_id: str, delay_ms: int) -> None:
+        """Schedule the next poll tick for a given group."""
+        session = self._sessions.get(group_id)
+        if not session:
+            return
+        delay = max(1, int(delay_ms))
+        self._scheduler.schedule(group_id, delay, lambda gid=group_id: self._on_poll_tick(gid))
+
+    def _on_poll_tick(self, group_id: str) -> None:
+        """Cooperative poll tick executed on the Tkinter thread."""
+        session = self._sessions.get(group_id)
+        if not session:
+            return
+        if not self._ensure_adapter():
+            return
+
+        coordinator = session.coordinator
+        context = session.context
+
+        tick: FlowTick = coordinator.poll_once(context)
+        if tick.event == "tick":
+            delay = tick.next_delay_ms
+            if delay is None:
+                base = getattr(self.settings_vm, "poll_interval_ms", 1000) or 1000
+                try:
+                    delay = max(200, int(base))
+                except (TypeError, ValueError):
+                    delay = 1000
+            self._schedule_poll(group_id, int(delay))
+            return
+
+        self._scheduler.cancel(group_id)
+        if tick.event == "completed":
+            coordinator.stop_polling()
+            auto_download_enabled = getattr(
+                self.settings_vm, "auto_download_on_complete", True
+            )
+            if tick.snapshot:
+                try:
+                    download_path = coordinator.on_completed(context, tick.snapshot)
+                except Exception as exc:
+                    self._toast_error(exc, context="Download failed")
+                    self._finalize_session(group_id)
+                    return
+                if auto_download_enabled:
+                    self._on_group_completed(group_id, download_path)
+            if not auto_download_enabled:
+                self._on_group_completed(group_id, None)
+            self._finalize_session(group_id)
+            return
+
+        if tick.event == "error":
+            coordinator.stop_polling()
+            self._finalize_session(group_id)
+
+    def _finalize_session(self, group_id: str) -> None:
+        """Clean up local bookkeeping once a run no longer needs polling."""
+        self._sessions.pop(group_id, None)
+        self._scheduler.cancel(group_id)
+        self.runs.unregister_runtime(group_id)
+        if self._active_group_id == group_id:
+            self._active_group_id = next(iter(self._sessions), None)
+            self.win.set_run_group_id(self._active_group_id or "")
+        self._refresh_runs_panel()
+
+    def _on_group_started(self, group_id: str, ctx: GroupContext) -> None:
+        self._log.debug("Coordinator acknowledged start for group %s", ctx.group)
+
+    def _on_group_snapshot(self, group_id: str, snapshot) -> None:
+        self.runs.update_snapshot(group_id, snapshot)
+        if snapshot and group_id == self._active_group_id:
+            self.progress_vm.apply_snapshot(snapshot)
+        self._refresh_runs_panel()
+
+    def _on_group_completed(self, group_id: str, path: Optional[Path]) -> None:
+        download_path = os.path.abspath(str(path)) if path else None
+        self.runs.mark_done(group_id, download_path)
+        if download_path:
+            self._last_download_dir = download_path
+            self.win.show_toast(self._build_download_toast(group_id, download_path))
+        else:
+            self.win.show_toast("All runs completed.")
+        self._refresh_runs_panel()
+        if getattr(self.progress_vm, "active_group_id", None) == group_id:
+            self.progress_vm.set_active_group(group_id, self.runs)
+
+    def _on_group_error(self, group_id: str, message: str) -> None:
+        text = message.strip() if isinstance(message, str) else ""
+        if text:
+            self._log.error("Polling error for %s: %s", group_id, text)
+        else:
+            self._log.error("Polling error for %s: <empty message>")
+        self.runs.mark_error(group_id, text or None)
+        self.win.show_toast(text or "Polling failed.")
+        self._refresh_runs_panel()
+
+    # ------------------------------------------------------------------
+    # Plan building
+    # ------------------------------------------------------------------
+    def _build_domain_plan(self) -> ExperimentPlan:
+        """Build a domain ExperimentPlan from the current view-model state."""
+        configured = self.plate_vm.configured()
+        if not configured:
+            raise RuntimeError("No configured wells to start.")
+
+        well_plan_list = self.experiment_vm.build_well_plan_map(configured)
+        if not well_plan_list:
+            raise RuntimeError("No saved parameters found for configured wells.")
+
+        inputs = self._collect_plan_inputs()
+        meta = build_meta(
+            experiment=inputs["experiment"],
+            subdir=inputs["subdir"],
+            client_dt_local=inputs["client_dt"],
+        )
+        self._last_plan_inputs = inputs
+
+        return ExperimentPlan(
+            meta=meta,
+            wells=well_plan_list,
+        )
+
+    def _collect_plan_inputs(self) -> Dict[str, Any]:
+        """Gather experiment metadata and filesystem settings for plan construction."""
+        experiment_name = (self.settings_vm.experiment_name or "").strip()
+        if not experiment_name:
+            raise RuntimeError("Experiment name must be set in Settings.")
+
+        subdir_raw = (self.settings_vm.subdir or "").strip()
+        subdir = subdir_raw or None
+
+        override_dt = ""
+        if hasattr(self.experiment_vm, "fields"):
+            override_dt = str(self.experiment_vm.fields.get("storage.client_datetime") or "")
+        client_dt = self._parse_client_datetime_override(override_dt)
+
+        results_dir = str(self.settings_vm.results_dir or "").strip() or "."
+
+        return {
+            "experiment": experiment_name,
+            "subdir": subdir,
+            "client_dt": client_dt,
+            "results_dir": results_dir,
+        }
+
+    def _parse_client_datetime_override(self, value: str) -> datetime:
+        """Interpret stored client datetime overrides, falling back to the current time."""
+        text = str(value or "").strip()
+        if not text:
+            return datetime.now().astimezone().replace(microsecond=0)
+
+        normalized = text.replace(" ", "T")
+        if normalized.endswith("Z"):
+            normalized = normalized[:-1] + "+00:00"
+        if "T" in normalized:
+            date_part, time_part = normalized.split("T", 1)
+            if ":" not in time_part:
+                time_part = time_part.replace("-", ":")
+            normalized = f"{date_part}T{time_part}"
+
+        parsed = None
+        try:
+            parsed = datetime.fromisoformat(normalized)
+        except ValueError:
+            pass
+
+        if parsed is None:
+            for fmt in ("%Y-%m-%d_%H-%M-%S", "%Y-%m-%d %H:%M:%S", "%Y-%m-%dT%H-%M-%S"):
+                try:
+                    parsed = datetime.strptime(text, fmt)
+                    break
+                except ValueError:
+                    continue
+
+        if parsed is None:
+            return datetime.now().astimezone().replace(microsecond=0)
+
+        if parsed.tzinfo is None or parsed.tzinfo.utcoffset(parsed) is None:
+            local_zone = datetime.now().astimezone().tzinfo
+            parsed = parsed.replace(tzinfo=local_zone)
+
+        return parsed.astimezone().replace(microsecond=0)
+
+    @staticmethod
+    def _format_client_datetime_for_storage(dt: datetime) -> str:
+        """Return a filesystem-safe representation of the client datetime."""
+        return dt.astimezone().strftime("%Y-%m-%d_%H-%M-%S")
+
+    # ------------------------------------------------------------------
+    # Download toast helpers
+    # ------------------------------------------------------------------
+    def _build_download_toast(self, group_id: str, path: str) -> str:
+        short_path = self._shorten_download_path(path)
+        descriptor = ""
+        meta = self._group_storage_meta.get(group_id) if group_id else None
+        if not meta and group_id:
+            entry = self.runs.get(group_id)
+            if entry:
+                meta = entry.storage_meta
+        if meta:
+            parts = [
+                str(meta.get("experiment") or "").strip(),
+                str(meta.get("subdir") or "").strip(),
+                str(meta.get("client_datetime") or "").strip(),
+            ]
+            descriptor = "/".join([p for p in parts if p])
+        target = f"{descriptor} -> {short_path}" if descriptor else short_path
+        if self._can_open_results_folder():
+            return f"Results unpacked to {target} (Ctrl+Shift+O to open)"
+        return f"Results unpacked to {target}"
+
+    @staticmethod
+    def _shorten_download_path(path: str, max_len: int = 60) -> str:
+        normalized = os.path.normpath(path)
+        if len(normalized) <= max_len:
+            return normalized
+        suffix_len = max(3, max_len - 3)
+        return f"...{normalized[-suffix_len:]}"
+
+    @staticmethod
+    def _can_open_results_folder() -> bool:
+        if sys.platform.startswith("win"):
+            return hasattr(os, "startfile")
+        if sys.platform == "darwin":
+            return True
+        return shutil.which("xdg-open") is not None
+
+    # ------------------------------------------------------------------
+    # Active group helpers
+    # ------------------------------------------------------------------
+    def _set_active_group(self, group_id: Optional[str]) -> None:
+        self._active_group_id = group_id
+        self.win.set_run_group_id(group_id or "")
+        self.runs_vm.set_active_group(group_id)
+        self.progress_vm.set_active_group(group_id, self.runs)
+        self._refresh_runs_panel()
+
+
+__all__ = ["FlowSession", "RunFlowPresenter"]
